+++ CHANGELOG.md
@@ -1,57 +1,80 @@
 ## NEXT
+
 - Adds `startWatchApp` method for iOS
 
+## 0.6.4
+
+- Upgrades `androidx.health:health-services-client` to `1.1.0-alpha03`
+- Default `WorkoutReading` timestamp is now `DateTime.timestamp()` instead of `DateTime.now()`
+
 ## 0.6.3
+
 - Upgrades permission_handler to version 11.0.0
 
 ## 0.6.2
+
 - Adds namespace to `build.gradle` for Gradle 8
 
 ## 0.6.1
+
 - Small documentation update [#5](https://github.com/Rexios80/flutter_workout/pull/5)
 
 ## 0.6.0
+
 - BREAKING: `ExerciseType` has been updated to match the native Android values
 - Updated Health Services. All data types now work on the Galaxy Watch4 and Pixel Watch.
 
 ## 0.5.1
+
 - Fixes Android build issue
 
 ## 0.5.0
+
 - BREAKING CHANGE: Updated Android compileSdkVersion to 33 for `permission_handler`
 - Updated dependencies
 
 ## 0.4.1
+
 - Fixed proguard issue
 
 ## 0.4.0
+
 - Support for new Health Services APIs
 
 ## 0.3.0
+
 - Updated Tizen code
 
 ## 0.2.1
+
 - Reverted gradle update for plugin
 
 ## 0.2.0
+
 - Updated linting rules
 - Updated gradle and target sdk
 - Improved package structure
 
 ## 0.1.0
+
 - Heart rate and calories works on Android. Others might work but don't show up in the emulator.
 
 ## 0.0.5
+
 - Run a foreground service on Android (untested)
 
 ## 0.0.4
+
 - Converted to Health Services for Wear OS (untested)
 
 ## 0.0.3
+
 - Changed com.example to dev.rexios
 
 ## 0.0.2
+
 - Added documentation
 
 ## 0.0.1
+
 - Initial release